+++ apps/commerce-streamer/src/main/java/com/loopers/application/support/cache/CacheEvictionService.java
@@ -1,4 +1,4 @@
-package com.loopers.application.service;
+package com.loopers.application.support.cache;
 
 import java.util.Set;
 import java.util.stream.Collectors;
